✨ Estimate nft tx gas according to size
