+++ src/transactions/gas.ts
@@ -1,6 +1,15 @@
 import { StdFee } from '@cosmjs/stargate';
 import BigNumber from 'bignumber.js';
-import { DEFAULT_GAS_PRICE_NUMBER, DEFAULT_MESSAGE_GAS } from '../constant';
+import jsonStringify from 'fast-json-stable-stringify';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { EncodeObject } from '@cosmjs/proto-signing';
+import {
+  GAS_ESTIMATOR_INTERCEPT,
+  GAS_ESTIMATOR_SLOPE,
+  GAS_ESTIMATOR_BUFFER_RATIO,
+  DEFAULT_GAS_PRICE_NUMBER,
+  DEFAULT_MESSAGE_GAS,
+} from '../constant';
 
 export function formatGasFee({
   gas = DEFAULT_MESSAGE_GAS,
@@ -26,4 +35,32 @@ export function formatGasFee({
   return fee;
 }
 
-export default formatGasFee;
+export function estimateMsgTxGas(msgs: EncodeObject[], {
+  denom,
+  gasPrice = DEFAULT_GAS_PRICE_NUMBER,
+  memo,
+}: {
+    denom: string,
+    gasPrice?: number,
+    memo?: string,
+  }): StdFee {
+  const value = {
+    msg: msgs,
+    // temp number here for estimation
+    fee: formatGasFee({ gas: '200000', gasPrice: '1', denom }),
+  };
+  const obj = {
+    type: 'cosmos-sdk/StdTx',
+    value,
+    memo, // directly append memo to object if exists, since we only need its length
+  };
+  const txBytes = Buffer.from(jsonStringify(obj), 'utf-8');
+  const byteSize = new BigNumber(txBytes.length);
+  const gasUsedEstimationBeforeBuffer = byteSize
+    .multipliedBy(GAS_ESTIMATOR_SLOPE)
+    .plus(GAS_ESTIMATOR_INTERCEPT);
+  const buffer = gasUsedEstimationBeforeBuffer.multipliedBy(GAS_ESTIMATOR_BUFFER_RATIO);
+  const gasUsedEstimation = gasUsedEstimationBeforeBuffer.plus(buffer);
+  const gas = gasUsedEstimation.toFixed(0, 0);
+  return formatGasFee({ gas, gasPrice, denom });
+}
